DEV-7652 Configuration into correct place
