+++ phpstan-baseline.neon
@@ -3,12 +3,12 @@ parameters:
 		-
 			message: "#^Call to an undefined method Symfony\\\\Component\\\\Config\\\\Definition\\\\Builder\\\\TreeBuilder\\:\\:root\\(\\)\\.$#"
 			count: 1
-			path: src/DependencyInjection/SkrzBunnyExtension.php
+			path: src/DependencyInjection/Configuration/Configuration.php
 
 		-
 			message: "#^Comparison operation \"\\<\" between 50307 and 40430 is always false\\.$#"
 			count: 1
-			path: src/DependencyInjection/SkrzBunnyExtension.php
+			path: src/DependencyInjection/Configuration/Configuration.php
 
 		-
 			message: "#^Method Skrz\\\\Bundle\\\\BunnyBundle\\\\Service\\\\BunnyManager\\:\\:__construct\\(\\) has parameter \\$config with no value type specified in iterable type array\\.$#"
